fixed bow 2 shot bug, made player rotate to location clicked before firing, fixed infinite companions bug,

+++ Assets/Development/Meyer/RedoInventory/UIItemsWithLabels.cs
@@ -50,7 +50,7 @@ public class UIItemsWithLabels : MonoBehaviour
     {
         for (int i = 0; i < Labels.Count; i++)
         {
-            Labels[i].labelText.text = stats[Labels[i].labelObject.name].ToString();
+            Labels[i].labelText.text = ((int)stats[Labels[i].labelObject.name]).ToString();
         }
     }
